+++ packages/react-components/react-teaching-popover-preview/src/components/TeachingPopoverCarouselNav/renderTeachingPopoverCarouselNav.tsx
@@ -1,0 +1,14 @@
+/** @jsxRuntime automatic */
+/** @jsxImportSource @fluentui/react-jsx-runtime */
+import type { TeachingPopoverCarouselNavState } from './TeachingPopoverCarouselNav.types';
+import { assertSlots } from '@fluentui/react-utilities';
+import { TeachingPopoverCarouselNavSlots } from './TeachingPopoverCarouselNav.types';
+
+/**
+ * Render the final JSX of TeachingPopoverCarouselNav
+ */
+export const renderTeachingPopoverCarouselNav_unstable = (state: TeachingPopoverCarouselNavState) => {
+  assertSlots<TeachingPopoverCarouselNavSlots>(state);
+
+  return <state.root />;
+};